Change pdf modifications handling
This change stores all modifications in a list and renders them on top
of the viewport, this lays the foundation for moving the modifications
after placement.
This replaces the old design of adding modifications to the document
itself and re-rendering it.

A positive side-effect of this is improved performance when adding a
modification. the previous design was re-rendering the entire document
on each modification.