Fix pdfjs error when deleting a label
The issue was caused by the delete action causing changes in two
different contexts (counter and modification) and each would cause
a re-render of [PdfViewportController -> PdfViewport -> PdfCanvas]
which caused invocation of [PdfCanvas.componentDidUpdate ->
renderPdfToCanvas].

As [renderPdfToCanvas] needs to be invoked only when page or scale
changes we can prevent any call to [renderPdfToCanvas] when there is
an unrelated update to PdfCanvas component.

This fix also improves performance by not re-rendering pdf to canvas
when there is no need to.